+++ packages/apollo/src/commands/service/__tests__/check.test.ts
@@ -30,9 +30,43 @@ describe("service:check", () => {
       ).toMatchInlineSnapshot(`
 "
 ### Apollo Service Check
-🔄 Validated your local schema against schema tag 'staging' on service 'engine'.
-🔢 Compared **18 schema changes** against operations seen over the **last 24 hours**.
-❌ Found **7 breaking changes** that would affect **3 operations**
+🔄 Validated your local schema against schema tag \`staging\` on service \`engine\`.
+🔢 Compared **18 schema changes** against **100 operations** seen over the **last 24 hours**.
+❌ Found **7 breaking changes** that would affect **3 operations** across **2 clients**
+
+🔗 [View your service check details](https://engine-dev.apollographql.com/service/engine/checks?schemaTag=Detached%3A%20d664f715645c5f0bb5ad4f2260cd6cb8d19bbc68&schemaTagId=f9f68e7e-1b5f-4eab-a3da-1fd8cd681111&from=2019-03-26T22%3A25%3A12.887Z).
+"
+`);
+      // Check when all the values are singluar
+      expect(
+        formatMarkdown({
+          serviceName: "engine",
+          tag: "staging",
+          checkSchemaResult: {
+            ...checkSchemaResult,
+            diffToPrevious: {
+              ...checkSchemaResult.diffToPrevious,
+              affectedClients: [
+                checkSchemaResult.diffToPrevious.affectedClients[0]
+              ],
+              affectedQueries: [
+                checkSchemaResult.diffToPrevious.affectedQueries[0]
+              ],
+              changes: [
+                checkSchemaResult.diffToPrevious.changes.find(
+                  change => change.type === ChangeType.FAILURE
+                )
+              ],
+              numberOfCheckedOperations: 1
+            }
+          }
+        })
+      ).toMatchInlineSnapshot(`
+"
+### Apollo Service Check
+🔄 Validated your local schema against schema tag \`staging\` on service \`engine\`.
+🔢 Compared **1 schema change** against **1 operation** seen over the **last 24 hours**.
+❌ Found **1 breaking change** that would affect **1 operation** across **1 client**
 
 🔗 [View your service check details](https://engine-dev.apollographql.com/service/engine/checks?schemaTag=Detached%3A%20d664f715645c5f0bb5ad4f2260cd6cb8d19bbc68&schemaTagId=f9f68e7e-1b5f-4eab-a3da-1fd8cd681111&from=2019-03-26T22%3A25%3A12.887Z).
 "
@@ -49,6 +83,7 @@ describe("service:check", () => {
             diffToPrevious: {
               ...checkSchemaResult.diffToPrevious,
               type: ChangeType.NOTICE,
+              affectedClients: [],
               affectedQueries: [],
               changes: []
             }
@@ -57,8 +92,8 @@ describe("service:check", () => {
       ).toMatchInlineSnapshot(`
 "
 ### Apollo Service Check
-🔄 Validated your local schema against schema tag 'staging' on service 'engine'.
-🔢 Compared **0 schema changes** against operations seen over the **last 24 hours**.
+🔄 Validated your local schema against schema tag \`staging\` on service \`engine\`.
+🔢 Compared **0 schema changes** against **100 operations** seen over the **last 24 hours**.
 ✅ Found **no breaking changes**.
 
 🔗 [View your service check details](https://engine-dev.apollographql.com/service/engine/checks?schemaTag=Detached%3A%20d664f715645c5f0bb5ad4f2260cd6cb8d19bbc68&schemaTagId=f9f68e7e-1b5f-4eab-a3da-1fd8cd681111&from=2019-03-26T22%3A25%3A12.887Z).